move OpenFileDialog to platform
