+++ NPLRuntime/Platform/iOS/src/ParaScriptBuildings/ParaScriptingGlobaliOS.cpp
@@ -1,0 +1,15 @@
+//-----------------------------------------------------------------------------
+// Class: For ParaGlobal namespace.
+// Authors: Big
+// Emails: onedous@gmail.com
+// Date: 2019.8.12
+//-----------------------------------------------------------------------------
+
+#include "ParaEngine.h"
+#include "ParaScriptingGlobal.h"
+#include <luabind/object.hpp>
+
+bool ParaScripting::ParaGlobal::OpenFileDialog(const object& inout)
+{
+	//refact to platform
+}
